remove preemptible from long steps of gnomix
