+++ docker/trtools-annotatr/Makefile
@@ -10,12 +10,8 @@ tag = ${version}--${git_commit}
 build:
 	# do the docker build
 	docker build --platform="linux/amd64" -t ${name}:${tag} .
-	# also build "native" (mostly for new macs...)
-	docker build -t ${name}:${tag}-native .
 	docker tag ${name}:${tag} ${name}:latest
-	docker tag ${name}:${tag}-native ${name}:latest-native
 
 push:
 	# Requires ~/.dockercfg
 	docker push ${name}:${tag}
-	docker push ${name}:latest-native
