+++ local_ancestry/README.md
@@ -37,6 +37,6 @@ cromshell list-outputs $jobid
 
 # Also keep track of useful phased Beagle files!
 chrom=chr11
-cromshell -mc -t 200 metadata $jobid > gnomix_${chrom}_metadata.json
-python -c "import json; data = json.load(open('gnomix_chr11_metadata.json', 'r')); [print(data['calls']['local_ancestry.run_gnomix'][i]['subWorkflowMetadata']['calls']['run_gnomix.beagle'][0]['outputs']['outvcf']) for i in range(len(data['calls']['local_ancestry.run_gnomix']))]; [print(data['calls']['local_ancestry.run_gnomix'][i]['subWorkflowMetadata']['calls']['run_gnomix.beagle'][0]['outputs']['outvcf_index']) for i in range(len(data['calls']['local_ancestry.run_gnomix']))];" | xargs -n 1 -P 1 -I% sh -c "gsutil mv % ${WORKSPACE_BUCKET}/beagle_hg19/${chrom}"
+cromshell -mc -t 2000 metadata $jobid > gnomix_${chrom}_metadata.json
+python -c "import json; data = json.load(open('gnomix_chr1_metadata.json', 'r')); [print(data['calls']['local_ancestry.run_gnomix'][i]['subWorkflowMetadata']['calls']['run_gnomix.beagle'][0]['outputs']['outvcf']) for i in range(len(data['calls']['local_ancestry.run_gnomix']))]; [print(data['calls']['local_ancestry.run_gnomix'][i]['subWorkflowMetadata']['calls']['run_gnomix.beagle'][0]['outputs']['outvcf_index']) for i in range(len(data['calls']['local_ancestry.run_gnomix']))];" | xargs -n 1 -P 1 -I% sh -c "echo gsutil mv % ${WORKSPACE_BUCKET}/beagle_hg19/${chrom}"
 ```
